requirements.txt
flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
python-dotenv==1.0.1
waitress==3.0.0